+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Timeline not showing events if the event occurs outside the `EXECUTION_STARTED` + `EXECUTION_FINISHED` events ([#180][#180])
 - Timeline incorrectly showing some `VF_APEX_CALL_START` events when dealing with ApexPage messages ([#212][#212])
+- Timeline tooltip not shown when zooming ([#242][#242])
 - Font sizes not correctly scaling in some places ([#238][#238])
 
 ## [1.5.2] - 2022-11-08
@@ -212,4 +213,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#219]: https://github.com/financialforcedev/debug-log-analyzer/issues/219
 [#129]: https://github.com/financialforcedev/debug-log-analyzer/issues/129
 [#238]: https://github.com/financialforcedev/debug-log-analyzer/issues/238
+[#242]: https://github.com/financialforcedev/debug-log-analyzer/issues/242
 [#235]: https://github.com/financialforcedev/debug-log-analyzer/issues/235